test: split military_units_panel_test_part5 to satisfy file-size lint
Move the two split-home-army widget tests and their _ArmySplitTestHarness
into a new part6 file so each file stays under the 1000 non-comment-line
repo-lint limit (part5 was 1024). Trim imports left unused after the move.

Refs #3514
